improved pattern expression parsing
