+++ bespokeasm/expression/__init__.py
@@ -15,8 +15,9 @@ import sys
 from bespokeasm.utilities import is_string_numeric, parse_numeric_string
 from bespokeasm.assembler.line_identifier import LineIdentifier
 from bespokeasm.assembler.label_scope import LabelScope
-from bespokeasm.assembler.line_object import is_valid_label
+from bespokeasm.assembler.line_object.utility import is_valid_label
 
+EXPRESSION_PARTS_PATTERN = r'(?:(?:\%|b)[01]+|(?:\$|0x)[0-9a-fA-F]+|\d+|[\+\-\*\/\&\|\^\(\)]|>>|<<|(?:\.|_)?\w+)'
 
 class TokenType(enum.Enum):
     T_NUM = 0
@@ -123,7 +124,7 @@ def _lexical_analysis(line_id: LineIdentifier, s: str) -> list[ExpressionNode]:
     }
 
     tokens = []
-    expression_parts = re.findall(r'(?:(?:\%|b)[01]+|(?:\$|0x)[0-9a-fA-F]+|\d+|[\+\-\*\/\&\|\^\(\)]|\>\>|\<\<|(?:\.|_)?\w+)', s)
+    expression_parts = re.findall(EXPRESSION_PARTS_PATTERN, s)
     for part in expression_parts:
         if part in mappings:
             token_type = mappings[part]
